# openapi.yaml
openapi: 3.0.3
info:
  title: Modullar Advancia Healthcare API
  description: Healthcare payment and compliance management platform
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        appointment_date:
          type: string
          format: date
        appointment_time:
          type: string
          format: time
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show, rescheduled]
        reason_for_visit:
          type: string
...
paths:
...
  /appointments:
    post:
      summary: Create appointment
      tags: [Appointments]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patient_id, provider_id, appointment_date, appointment_time]
              properties:
                patient_id:
                  type: string
                  format: uuid
                provider_id:
                  type: string
                  format: uuid
                appointment_date:
                  type: string
                  format: date
                appointment_time:
                  type: string
                  format: time
                reason_for_visit:
                  type: string
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
...
  /disputes:
    get:
      summary: List all disputes
      tags: [Disputes]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of disputes
    post:
      summary: Create dispute
      tags: [Disputes]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dispute_number, dispute_reason, amount]
              properties:
                dispute_number:
                  type: string
                transaction_id:
                  type: string
                  format: uuid
                dispute_reason:
                  type: string
                  enum: [fraud, duplicate, product_not_received, service_not_provided, other]
                amount:
                  type: number
                customer_description:
                  type: string
      responses:
        '201':
          description: Dispute created
  /webhooks:
    get:
      summary: List user webhooks
      tags: [Webhooks]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of webhooks
    post:
      summary: Create webhook
      tags: [Webhooks]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, url, secret_key]
              properties:
                name:
                  type: string
                url:
                  type: string
                  format: uri
                secret_key:
                  type: string
                subscribed_events:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Webhook created
...